requirements.txt
python-telegram-bot==22.3
pydantic==2.11.7
pydantic-settings
git+https://github.com/dronnikovigor/yandex-music-downloader.git@develop#egg=yandex-music-downloader-api
